+++ external/SLikeNet/Source/src/DynDNS.cpp
@@ -7,7 +7,7 @@
  *  of patent rights can be found in the RakNet Patents.txt file in the same directory.
  *
  *
- *  Modified work: Copyright (c) 2016-2017, SLikeSoft UG (haftungsbeschr‰nkt)
+ *  Modified work: Copyright (c) 2016-2018, SLikeSoft UG (haftungsbeschr√§nkt)
  *
  *  This source code was modified by SLikeSoft. Modifications are licensed under the MIT-style
  *  license found in the license.txt file in the root directory of this source tree.
@@ -171,7 +171,7 @@ void DynDNS::Update(void)
 							{
 								SystemAddress parser;
 								parser.FromString(curResult);
-								parser.ToString(false, myIPStr, 32);
+								parser.ToString(false, myIPStr, static_cast<size_t>(32));
 							}
 						}
 						tcp->DeallocatePacket(packet);
@@ -218,7 +218,7 @@ void DynDNS::Update(void)
 				curResult+=strlen("Current IP Address: ");
 				SystemAddress myIp;
 				myIp.FromString(curResult);
-				myIp.ToString(false, myIPStr, 32);
+				myIp.ToString(false, myIPStr, static_cast<size_t>(32));
 
 				char existingHost[65];
 				existingHost[0]=0;
