+++ external/SLikeNet/Source/src/NatTypeDetectionServer.cpp
@@ -7,7 +7,7 @@
  *  of patent rights can be found in the RakNet Patents.txt file in the same directory.
  *
  *
- *  Modified work: Copyright (c) 2016-2017, SLikeSoft UG (haftungsbeschr‰nkt)
+ *  Modified work: Copyright (c) 2016-2018, SLikeSoft UG (haftungsbeschr√§nkt)
  *
  *  This source code was modified by SLikeSoft. Modifications are licensed under the MIT-style
  *  license found in the license.txt file in the root directory of this source tree.
@@ -54,7 +54,7 @@ void NatTypeDetectionServer::Startup(
 	DataStructures::List<RakNetSocket2* > sockets;
 	rakPeerInterface->GetSockets(sockets);
 	char str[64];
-	sockets[0]->GetBoundAddress().ToString(false,str,64);
+	sockets[0]->GetBoundAddress().ToString(false,str,static_cast<size_t>(64));
 	s1p2=
 		CreateNonblockingBoundSocket(str,
 #ifdef __native_client__
